Working on the interpreter and database
