+++ README.md
@@ -3,10 +3,10 @@ Another Smart Meter Project, which allows users to read the values of their (dut
 
 Milestones:
 
-1. Getting to know git better
-2. Creating a hello world python application
+1. (Done)Getting to know git better
+2. (Done)Creating a hello world python application
 3. Setting up a docker scipt file to contain the application
-4. Read values from the p1 port and display the result in the console
+4. (Done)Read values from the p1 port and display the result in the console
 5. Store values in the database
 6. Create a web application to display the data from the database
 7. Create a fancy web application
